// clients/contentService/models/WebPageCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WebPageCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    tenantId?: string | null;
    enrolmentId?: string | null;
    title?: string | null;
    authorId?: string | null;
    description?: string | null;
    htmlContent?: string | null;
    featuredImageUrl?: string | null;
    webTemplateID?: string | null;
};